Break down file_store.py into separate files
for interfaces and implementations
